sensor-data-pipeline/requirements.txt
kafka-python
tb-mqtt-client
faust
typing-extensions
numpy
cassandra-driver
mysql-connector-python
opentelemetry-distro
opentelemetry-instrumentation
opentelemetry-exporter-otlp